requirements.txt
# StudyWiseAI Backend Dependencies

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
# psycopg2-binary==2.9.9  # For production PostgreSQL
# redis==5.0.1  # Optional for production caching

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & ML Libraries  
openai==1.3.5
# scikit-learn==1.3.2  # Commented out - requires Visual C++ Build Tools on Windows  
pandas==2.1.4
numpy==1.25.2

# HTTP Client
httpx==0.25.2

# Environment Variables
python-decouple==3.8

# Email Support
fastapi-mail==1.4.1

# WebSocket Support
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS
python-cors==0.1.0

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Background Tasks
# celery==5.3.4  # Optional for production

# Development
black==23.11.0
isort==5.12.0